requirements.txt
bleak
numpy
Pillow
pygame-ce
